changed spacing in tranposed p1000 tip rack
The spacing has been changed so that the first tip to be picked up can now be in the A1 position of the tip rack and the next one will be in A2 position, i.e. one step closer to the front edge of the deck. Although the previous spacing is consistent with the general layout of the deck with row A closest to the front edge, it seems more intuitive to have the first tip picked up from the top left (A1) position of a tip rack.